Query engine reorg
There were a couple unsatisfactory elements of the query engine.

- Public fields which goes against C-STRUCT-PRIVATE
- Switch to computing cached data on demand instead of pre-emptively.
This will allow eu4save to start caching a lot of computations. One of
this is for calculating the timeline of which buildings have been built,
which requires looking at all the provinces. If someone never looks at
building info, then they shouldn't need to pay that caching operation
cost.