Rework CountryTag to validate inline storage
Making CountryTag a newtype wrapper around String meant that it was
large (24 bytes + whatever on the heap). Since a country tag is only
three bytes long we can witness a 8x space savings by storing the three
bytes directly in the structure. This also makes CountryTag as Copyable
and more ergonomic to pass around.

In order to ensure invariants (that the country tag contains three bytes
that are alphanumeric or dashes) a country tag now validates data. This
has an ergonomic cost of needing to parse data into a country tag even
when the tag is constructed from a byte literal. We'll see how annoying
this turns out to be.