Add initial EU4 save formatter experiment
This adds a `fmt` command that takes a plaintext, uncompressed save and
formats into a ouput that can be more easily diffed with another save.

The output is not a valid save.

There can be additional improvements that revolve around allowing quotes
/ non-quotes to be included in the output. This is just the MVP that
found a few issues with accidental dates and 64 bit precision.