Melt invalid dates in non-error mode
When a strange binary date is encountered (like prior to 100 BC and more
recently prior to 5000 BC), unless the melter is in error mode, write
out the raw integer value instead of returning an error. This way the
melter can still output text that can be used by the game.