Allow customizing how tokens are resolved
Previously one had to specify the tokens through environment variable
pointing to a defined format file. This is not a great idea as there is
no customization possible. Instead this PR allows one to pass in how
tokens are resolved, while keeping it backwards compatible for those
using the environment variable.