Allow one pass `Eu4Save` deserialization
Previously it required a deserializer that wasn't mutable.

I don't want to have to parse to a tape or parse the input twice for
plaintext uncompressed EU4 saves.

The downside of copying Meta and Gamestate fields into a manually
flatten struct is the cumbersome issue of keeping the fields in sync and
the growth in executable size (pdx.tools recorded a 138 kB increase in
wasm size)